refactor: internal package to db for consistancy with account subgraph
